MFC r282881: Do not promote large async writes to sync.
Present implementation of large sync writes is too strict and so can be
quite slow.  Instead of doing that, execute large async write in chunks,
syncing each chunk separately.

It would be good to fix large sync writes too, but I leave it to somebody
with more skills in this area.


git-svn-id: https://svn.freebsd.org/base/stable/10@283457 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f